#10072: Merge asinh_bw to TTNN
