+++ docs/source/ttnn/ttnn/ttnn/asinh_bw.rst
@@ -1,0 +1,6 @@
+ _ttnn.asinh_bw:
+
+ttnn.asinh_bw
+##############
+
+  .. autofunction:: ttnn.asinh_bw
