+++ docs/source/ttnn/ttnn/ttnn/hardswish_bw.rst
@@ -1,6 +1,6 @@
 .. _ttnn.hardswish_bw:
 
 ttnn.hardswish_bw
-###############
+##################
 
   .. autofunction:: ttnn.hardswish_bw
