+++ docs/source/ttnn/ttnn/ttnn/tanhshrink_bw.rst
@@ -1,6 +1,6 @@
 .. _ttnn.tanhshrink_bw:
 
 ttnn.tanhshrink_bw
-###############
+###################
 
   .. autofunction:: ttnn.tanhshrink_bw
