+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -58,7 +58,8 @@ enum class UnaryBackwardOpType {
     HARDSWISH_BW,
     TANHSHRINK_BW,
     ATANH_BW,
-    ASIN_BW
+    ASIN_BW,
+    ASINH_BW
 };
 
 struct UnaryBackwardFunction{
